third_party/blink/web_tests/virtual/old-text-size-adjust/wpt_internal/css/css-size-adjust/text-size-adjust-affects-line-height-px-large-expected.txt

This is a testharness.js-based test.
[FAIL] Resolved line-height with specified 50px line-height and text-size-adjust: 150%
  assert_equals: expected "75px" but got "50px"
[FAIL] Actual line-height with specified 50px line-height and text-size-adjust: 150%
  assert_equals: expected 75 but got 50
Harness: the test ran to completion.
